requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
anytree==2.12.1
asttokens==2.4.1
attrs==25.1.0
autograd==1.7.0
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==4.9.0
azure-identity==1.20.0
azure-search-documents==11.5.2
azure-storage-blob==12.24.1
beartype==0.18.5
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
devtools==0.12.2
distro==1.9.0
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0.tar.gz#sha256=504fe70715dca9464bf11e385fbe80e92ae127394df3f9c45242926082425551
environs==11.2.1
executing==2.2.0
fastapi==0.115.8
fnllm==0.0.10
fonttools==4.56.0
frozenlist==1.5.0
future==1.0.0
gensim==4.3.3
geographiclib==2.0
geopy==2.4.1
graphrag==1.2.0
graspologic==3.4.1
graspologic-native==1.2.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyppo==0.4.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
isodate==0.7.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json_repair==0.30.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lancedb==0.17.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.10
language_data==1.3.0
lazy-object-proxy==1.10.0
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.6.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
murmurhash==1.0.12
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
openai==1.64.0
openapi-core==0.19.4
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
openrouteservice==2.3.3
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parse==1.20.2
pathable==0.4.4
patsy==1.0.1
pillow==11.1.0
pinecone==6.0.1
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
portalocker==2.10.1
POT==0.9.5
prance==23.6.21.0
preshed==3.0.9
propcache==0.3.0
pyarrow==15.0.2
pybars4==0.9.13
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pylance==0.20.0
PyMeta3==0.5.1
pynndescent==0.5.13
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
semantic-kernel==1.20.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
starlette==0.45.3
statsmodels==0.14.4
tenacity==9.0.0
textblob==0.18.0.post0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
